requirements.txt
# Core Flask
Flask==2.3.3
Flask-Login==0.6.2
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2

# Database
SQLAlchemy==2.0.23         # Optional – if you switch to ORM later
psycopg2-binary==2.9.9     # PostgreSQL adapter (for production)
# SQLite is built-in – no package needed

# Forms & Validation
WTForms==3.1.1
Flask-WTF==1.2.1

# Email
Flask-Mail==0.9.1

# File processing (bulk upload)
pandas==2.1.4
openpyxl==3.1.2            # Excel support for pandas

# PDF generation
reportlab==4.2.0

# Environment variables
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Security (optional but recommended)
Flask-Talisman==1.0.0       # Security headers
Flask-Limiter==3.5.0        # Rate limiting (protects against abuse)

# Monitoring / Debug (optional)
blinker==1.7.0              # Signal support (used by Flask-Limiter)

# Caching (optional – if you add Redis later)

redis==5.0.1
Pillow==10.0.0
qrcode==7.4.2
reportlab==4.0.4
python-barcode==0.14.0